refactor: rename supabase client module
