+++ src/contexts/AuthContext.tsx
@@ -1,6 +1,6 @@
 // src/contexts/AuthContext.tsx
 import { createContext, useContext, useEffect, useState } from 'react';
-import { supabase } from '../lib/supabaseClient';
+import { supabase } from '../lib/supabase';
 import type { Session, User, AuthChangeEvent } from '@supabase/supabase-js';
 
 interface AuthContextType {
